+++ packages/server/core/src/base/adapters/node/loadServer.ts
@@ -1,6 +1,6 @@
 import path from 'path';
 import { fs, dotenv, dotenvExpand } from '@modern-js/utils';
-import { ServerBaseOptions } from '../../../core/server';
+import type { ServerBaseOptions } from '../../serverBase';
 
 /** 读取 .env.{process.env.MODERN_ENV} 文件，加载环境变量 */
 export async function loadServerEnv(options: ServerBaseOptions) {
