+++ packages/server/core/src/base/adapters/node/middlewares/templates.ts
@@ -1,7 +1,7 @@
 import path from 'path';
 import { ServerRoute } from '@modern-js/types';
 import { fileReader } from '@modern-js/runtime-utils/fileReader';
-import { HonoMiddleware, ServerEnv } from '../../../../core/server';
+import { Middleware, ServerEnv } from '../../../../core/server';
 
 export async function getHtmlTemplates(pwd: string, routes: ServerRoute[]) {
   const htmls = await Promise.all(
@@ -16,7 +16,7 @@ export async function getHtmlTemplates(pwd: string, routes: ServerRoute[]) {
       return [route.entryName!, html];
     }) || [],
   );
-  // eslint-disable-next-line node/no-unsupported-features/es-builtins
+
   const templates: Record<string, string> = Object.fromEntries(htmls);
 
   return templates;
@@ -25,7 +25,7 @@ export async function getHtmlTemplates(pwd: string, routes: ServerRoute[]) {
 export function injectTemplates(
   pwd: string,
   routes?: ServerRoute[],
-): HonoMiddleware<ServerEnv> {
+): Middleware<ServerEnv> {
   return async (c, next) => {
     if (routes && !c.get('templates')) {
       const templates = await getHtmlTemplates(pwd, routes);
