+++ packages/server/core/src/base/middlewares/monitor.ts
@@ -1,7 +1,7 @@
 import type { Logger, Metrics, Reporter } from '@modern-js/types';
 import { time } from '@modern-js/runtime-utils/time';
 import { ServerReportTimings } from '../constants';
-import type { HonoContext, Next, ServerEnv } from '../../core/server';
+import type { Context, Next, ServerEnv } from '../../core/server';
 
 const defaultReporter: Reporter = {
   init() {
@@ -23,7 +23,7 @@ const defaultReporter: Reporter = {
 
 // TODO: unify
 export function injectReporter() {
-  return async (c: HonoContext<ServerEnv>, next: Next) => {
+  return async (c: Context<ServerEnv>, next: Next) => {
     const reporter = c.get('reporter');
     if (!reporter) {
       c.set('reporter', defaultReporter);
@@ -33,7 +33,7 @@ export function injectReporter() {
 }
 
 export function initReporter(entryName: string) {
-  return async (c: HonoContext<ServerEnv>, next: Next) => {
+  return async (c: Context<ServerEnv>, next: Next) => {
     const reporter = c.get('reporter');
 
     if (!reporter) {
@@ -54,7 +54,7 @@ export function initReporter(entryName: string) {
 }
 
 export function injectLogger(inputLogger: Logger) {
-  return async (c: HonoContext<ServerEnv>, next: Next) => {
+  return async (c: Context<ServerEnv>, next: Next) => {
     const logger = c.get('logger');
     if (!logger && inputLogger) {
       c.set('logger', inputLogger);
@@ -64,7 +64,7 @@ export function injectLogger(inputLogger: Logger) {
 }
 
 export function injectMetrics(inputMetrics: Metrics) {
-  return async (c: HonoContext<ServerEnv>, next: Next) => {
+  return async (c: Context<ServerEnv>, next: Next) => {
     const metrics = c.get('metrics');
 
     if (!metrics) {
