--- packages/server/core/src/core/hono.ts
@@ -1,7 +1,0 @@
-export type {
-  Context as HonoContext,
-  Env as HonoEnv,
-  HonoRequest,
-  Next,
-  MiddlewareHandler as HonoMiddleware,
-} from 'hono';
