+++ packages/solutions/app-tools/src/types/index.ts
@@ -14,9 +14,7 @@ export type { webpack, Rspack } from '@modern-js/uni-builder';
 export type { Bundler } from './utils';
 export type {
   ServerUserConfig,
-  ServerNormalizedConfig,
   BffUserConfig,
-  BffNormalizedConfig,
   SSR,
   SSRByEntries,
 } from '@modern-js/server-core';
